backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyGithub
langchain-text-splitters==0.2.4
chromadb==0.5.23
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-google-genai==1.0.10
google-generativeai==0.7.2
filetype
langchain-groq
groq
langchain-chroma==0.1.4
tokenizers==0.15.2
sentence-transformers 
cohere 
langchain-cohere 
cohere 
langchain-cohere 
